backend/requirements.txt
Flask
Pillow
imageio
pytest
# Add other dependencies as needed
